refactor(fleet-change): extract ship selection
