+++ autowsgr/ui/battle/fleet_change/__init__.py
@@ -8,7 +8,8 @@
 - ``_detect.py`` -- 准备页舰队 OCR 检测
 - ``_rules.py`` -- 舰队规则匹配与校验
 - ``_planning.py`` -- 目标分配与槽位规划
-- ``_alignment.py`` -- 舰队调整与船池页面操作
+- ``_selection.py`` -- 船池选择页面操作
+- ``_alignment.py`` -- 舰队成员调整与排序
 - ``_change.py`` -- 换船主流程编排
 """
